+++ src/main/java/faang/school/postservice/controller/PostController.java
@@ -4,7 +4,7 @@ import faang.school.postservice.dto.PostDto;
 import faang.school.postservice.dto.ResourceDto;
 import faang.school.postservice.exception.DataValidationException;
 import faang.school.postservice.exception.MaxUploadCountExceededException;
-import faang.school.postservice.service.PostService;
+import faang.school.postservice.service.PostServiceImpl;
 import jakarta.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -31,7 +31,7 @@ import java.util.List;
 @Slf4j
 public class PostController {
 
-    private final PostService postService;
+    private final PostServiceImpl postService;
 
     @PostMapping
     public PostDto createDraft(@RequestBody PostDto postDto) {
